refactor: move output target to its own dir
